backend/requirements.txt
flask
flask-cors
kafka-python
python-dotenv
requests
pinecone-client
sentence-transformers
langchain-groq
langchain-community
langdetect